Rename engine/logging.py to engine/config.py
